# backend/internal/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Defines values for PlanUnit.
const (
	Count    PlanUnit = "count"
	Distance PlanUnit = "distance"
	Time     PlanUnit = "time"
)

// Defines values for StatisticGroupBy.
const (
	Day   StatisticGroupBy = "day"
	Month StatisticGroupBy = "month"
	Year  StatisticGroupBy = "year"
)

// Habit defines model for Habit.
type Habit struct {
	CurrentPlan   *Plan               `json:"currentPlan,omitempty"`
	Id            *int                `json:"id,omitempty"`
	Name          *string             `json:"name,omitempty"`
	Notifications *bool               `json:"notifications,omitempty"`
	StartDate     *openapi_types.Date `json:"startDate,omitempty"`

	// TodayValue today in sec
	TodayValue *float32 `json:"todayValue,omitempty"`
}

// HabitStatisticResponse defines model for HabitStatisticResponse.
type HabitStatisticResponse struct {
	GroupBy *StatisticGroupBy `json:"groupBy,omitempty"`
	Period  *[]struct {
		// Interval тут приходит в зависимости от типа, условно если мы выбрали день, то приходят дни недели, если месяц, то месяца, аналогично с годами
		Interval *string `json:"interval,omitempty"`

		// Value тут приходит значение в зависимости от типа
		Value *int `json:"value,omitempty"`
	} `json:"period,omitempty"`
	PlanUnit *PlanUnit `json:"planUnit,omitempty"`
}

// HabitUpdate defines model for HabitUpdate.
type HabitUpdate struct {
	Description   *string `json:"description,omitempty"`
	Name          *string `json:"name,omitempty"`
	Notifications *bool   `json:"notifications,omitempty"`
	Plan          *Plan   `json:"plan,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	Token *string `json:"token,omitempty"`
}

// NewHabit defines model for NewHabit.
type NewHabit struct {
	Description   *string `json:"description,omitempty"`
	Name          string  `json:"name"`
	Notifications *bool   `json:"notifications,omitempty"`
	Plan          Plan    `json:"plan"`
}

// Plan defines model for Plan.
type Plan struct {
	// Goal тут крч мы передаем значение в зависимости от плана, в минимальных единицах измерения, (метры, секунды, разы)
	Goal     *int      `json:"goal,omitempty"`
	PlanUnit *PlanUnit `json:"planUnit,omitempty"`
}

// PlanUnit defines model for PlanUnit.
type PlanUnit string

// RegisterRequest defines model for RegisterRequest.
type RegisterRequest struct {
	Age      *int                 `json:"age,omitempty"`
	Email    *openapi_types.Email `json:"email,omitempty"`
	Password *string              `json:"password,omitempty"`
	Username *string              `json:"username,omitempty"`
}

// ScoreHabit defines model for ScoreHabit.
type ScoreHabit struct {
	Date *time.Time `json:"date,omitempty"`

	// Value тут крч мы передаем значение в минимальных единицах измерения, (метры, секунды, разы)
	Value *int `json:"value,omitempty"`
}

// StatisticGroupBy defines model for StatisticGroupBy.
type StatisticGroupBy string

// UserProfile defines model for UserProfile.
type UserProfile struct {
	AvatarUrl   *string             `json:"avatarUrl,omitempty"`
	City        *string             `json:"city,omitempty"`
	Country     *string             `json:"country,omitempty"`
	DateOfBirth *openapi_types.Date `json:"dateOfBirth,omitempty"`
	Email       *string             `json:"email,omitempty"`
	Username    *string             `json:"username,omitempty"`
}

// UserProfileUpdate defines model for UserProfileUpdate.
type UserProfileUpdate struct {
	City        *string             `json:"city,omitempty"`
	Country     *string             `json:"country,omitempty"`
	DateOfBirth *openapi_types.Date `json:"dateOfBirth,omitempty"`
	Email       *string             `json:"email,omitempty"`
	Name        *string             `json:"name,omitempty"`
	Password    *string             `json:"password,omitempty"`
	Username    *string             `json:"username,omitempty"`
}

// GetApiHabitsHabitIdStatisticParams defines parameters for GetApiHabitsHabitIdStatistic.
type GetApiHabitsHabitIdStatisticParams struct {
	DateFrom openapi_types.Date `form:"date-from" json:"date-from"`
	DateTo   openapi_types.Date `form:"date-to" json:"date-to"`
	GroupBy  StatisticGroupBy   `form:"group-by" json:"group-by"`
}

// PostApiUsersAvatarMultipartBody defines parameters for PostApiUsersAvatar.
type PostApiUsersAvatarMultipartBody struct {
	File *openapi_types.File `json:"file,omitempty"`
}

// PostApiAuthLoginJSONRequestBody defines body for PostApiAuthLogin for application/json ContentType.
type PostApiAuthLoginJSONRequestBody = LoginRequest

// PostApiAuthRegisterJSONRequestBody defines body for PostApiAuthRegister for application/json ContentType.
type PostApiAuthRegisterJSONRequestBody = RegisterRequest

// PostApiHabitsJSONRequestBody defines body for PostApiHabits for application/json ContentType.
type PostApiHabitsJSONRequestBody = NewHabit

// PutApiHabitsHabitIdJSONRequestBody defines body for PutApiHabitsHabitId for application/json ContentType.
type PutApiHabitsHabitIdJSONRequestBody = HabitUpdate

// PostApiHabitsHabitIdScoreJSONRequestBody defines body for PostApiHabitsHabitIdScore for application/json ContentType.
type PostApiHabitsHabitIdScoreJSONRequestBody = ScoreHabit

// PostApiUsersAvatarMultipartRequestBody defines body for PostApiUsersAvatar for multipart/form-data ContentType.
type PostApiUsersAvatarMultipartRequestBody PostApiUsersAvatarMultipartBody

// PutApiUsersProfileJSONRequestBody defines body for PutApiUsersProfile for application/json ContentType.
type PutApiUsersProfileJSONRequestBody = UserProfileUpdate

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Вход пользователя
	// (POST /api/auth/login)
	PostApiAuthLogin(ctx echo.Context) error
	// Регистрация нового пользователя
	// (POST /api/auth/register)
	PostApiAuthRegister(ctx echo.Context) error
	// Получить список хобби пользователя
	// (GET /api/habits)
	GetApiHabits(ctx echo.Context) error
	// Создать новое хобби
	// (POST /api/habits)
	PostApiHabits(ctx echo.Context) error
	// Получить детали хобби
	// (GET /api/habits/{habitId})
	GetApiHabitsHabitId(ctx echo.Context, habitId int) error
	// Обновить хобби
	// (PUT /api/habits/{habitId})
	PutApiHabitsHabitId(ctx echo.Context, habitId int) error
	// Записать время, потраченное на хобби
	// (POST /api/habits/{habitId}/score)
	PostApiHabitsHabitIdScore(ctx echo.Context, habitId int) error
	// Получение статустики
	// (GET /api/habits/{habitId}/statistic)
	GetApiHabitsHabitIdStatistic(ctx echo.Context, habitId int, params GetApiHabitsHabitIdStatisticParams) error
	// Получение статустики
	// (GET /api/habits/{habitId}/statistic/total)
	GetApiHabitsHabitIdStatisticTotal(ctx echo.Context, habitId int) error
	// Загрузить аватар
	// (POST /api/users/avatar)
	PostApiUsersAvatar(ctx echo.Context) error
	// Получить профиль пользователя
	// (GET /api/users/profile)
	GetApiUsersProfile(ctx echo.Context) error
	// Обновить профиль пользователя
	// (PUT /api/users/profile)
	PutApiUsersProfile(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostApiAuthLogin converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiAuthLogin(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiAuthLogin(ctx)
	return err
}

// PostApiAuthRegister converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiAuthRegister(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiAuthRegister(ctx)
	return err
}

// GetApiHabits converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiHabits(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiHabits(ctx)
	return err
}

// PostApiHabits converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiHabits(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiHabits(ctx)
	return err
}

// GetApiHabitsHabitId converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiHabitsHabitId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "habitId" -------------
	var habitId int

	err = runtime.BindStyledParameterWithOptions("simple", "habitId", ctx.Param("habitId"), &habitId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter habitId: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiHabitsHabitId(ctx, habitId)
	return err
}

// PutApiHabitsHabitId converts echo context to params.
func (w *ServerInterfaceWrapper) PutApiHabitsHabitId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "habitId" -------------
	var habitId int

	err = runtime.BindStyledParameterWithOptions("simple", "habitId", ctx.Param("habitId"), &habitId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter habitId: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutApiHabitsHabitId(ctx, habitId)
	return err
}

// PostApiHabitsHabitIdScore converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiHabitsHabitIdScore(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "habitId" -------------
	var habitId int

	err = runtime.BindStyledParameterWithOptions("simple", "habitId", ctx.Param("habitId"), &habitId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter habitId: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiHabitsHabitIdScore(ctx, habitId)
	return err
}

// GetApiHabitsHabitIdStatistic converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiHabitsHabitIdStatistic(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "habitId" -------------
	var habitId int

	err = runtime.BindStyledParameterWithOptions("simple", "habitId", ctx.Param("habitId"), &habitId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter habitId: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetApiHabitsHabitIdStatisticParams
	// ------------- Required query parameter "date-from" -------------

	err = runtime.BindQueryParameter("form", true, true, "date-from", ctx.QueryParams(), &params.DateFrom)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter date-from: %s", err))
	}

	// ------------- Required query parameter "date-to" -------------

	err = runtime.BindQueryParameter("form", true, true, "date-to", ctx.QueryParams(), &params.DateTo)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter date-to: %s", err))
	}

	// ------------- Required query parameter "group-by" -------------

	err = runtime.BindQueryParameter("form", true, true, "group-by", ctx.QueryParams(), &params.GroupBy)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter group-by: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiHabitsHabitIdStatistic(ctx, habitId, params)
	return err
}

// GetApiHabitsHabitIdStatisticTotal converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiHabitsHabitIdStatisticTotal(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "habitId" -------------
	var habitId int

	err = runtime.BindStyledParameterWithOptions("simple", "habitId", ctx.Param("habitId"), &habitId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter habitId: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiHabitsHabitIdStatisticTotal(ctx, habitId)
	return err
}

// PostApiUsersAvatar converts echo context to params.
func (w *ServerInterfaceWrapper) PostApiUsersAvatar(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostApiUsersAvatar(ctx)
	return err
}

// GetApiUsersProfile converts echo context to params.
func (w *ServerInterfaceWrapper) GetApiUsersProfile(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetApiUsersProfile(ctx)
	return err
}

// PutApiUsersProfile converts echo context to params.
func (w *ServerInterfaceWrapper) PutApiUsersProfile(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutApiUsersProfile(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/api/auth/login", wrapper.PostApiAuthLogin)
	router.POST(baseURL+"/api/auth/register", wrapper.PostApiAuthRegister)
	router.GET(baseURL+"/api/habits", wrapper.GetApiHabits)
	router.POST(baseURL+"/api/habits", wrapper.PostApiHabits)
	router.GET(baseURL+"/api/habits/:habitId", wrapper.GetApiHabitsHabitId)
	router.PUT(baseURL+"/api/habits/:habitId", wrapper.PutApiHabitsHabitId)
	router.POST(baseURL+"/api/habits/:habitId/score", wrapper.PostApiHabitsHabitIdScore)
	router.GET(baseURL+"/api/habits/:habitId/statistic", wrapper.GetApiHabitsHabitIdStatistic)
	router.GET(baseURL+"/api/habits/:habitId/statistic/total", wrapper.GetApiHabitsHabitIdStatisticTotal)
	router.POST(baseURL+"/api/users/avatar", wrapper.PostApiUsersAvatar)
	router.GET(baseURL+"/api/users/profile", wrapper.GetApiUsersProfile)
	router.PUT(baseURL+"/api/users/profile", wrapper.PutApiUsersProfile)

}